requirements.txt
psutil==5.9.5
fastapi==0.88.0
python-dotenv==0.21.0
pydantic==1.10.4
python-ms-core==0.0.16
uvicorn==0.20.0